Move BitMexPriceEventId to daemon::olivia
The `BitMexPriceEventId` is part of Olivia's API.
