+++ daemon/src/rollover_taker.rs
@@ -4,11 +4,11 @@ use crate::model::cfd::Dlc;
 use crate::model::cfd::OrderId;
 use crate::model::cfd::Role;
 use crate::model::cfd::RolloverCompleted;
-use crate::model::BitMexPriceEventId;
 use crate::model::FundingFee;
 use crate::model::FundingRate;
 use crate::model::Timestamp;
 use crate::model::TxFeeRate;
+use crate::olivia::BitMexPriceEventId;
 use crate::oracle;
 use crate::oracle::GetAnnouncement;
 use crate::process_manager;
